Automatically determine the network API port from GSheets
Some servers in our Lab have interface names that are different
then other servers. For example "eno1" may not exist on some
servers. Thus, the first interface with a carrier signal may not
be the correct interface. In our GSheet we should make sure that
the provisioning interface is aligned with the Provision Host
column.

Signed-off-by: William Zhao <wizhao@redhat.com>