Prevent stp to cause interface going down
When one creates the virbr0 and attaches the physical interface to it,
by default, some spanning-tree messages are generated.
Some switches, using default configuration,  do not expect such messages
and disable the interface - this can be seen on the localhost
where the bridge is running with "ethtool <ifname>", showing interface
disabled.
Disable stp on the virbr0, so the patch does not have to deal with switches
configuration

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>