Added option to install iso on bm hosting vms and boot those bms
If <pre-installed> is set to False, we will use a fcos iso build
with additional packages such as libvirt and boot bm from it.
THis could be useful if one wants to do some quick test without
having to pre-install bm systems.
bm installed this way do not survide reboot.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>