Wait for each sriov resource to be applied
On occasion an edge case can be reached where the initial wait for the
sriov mcp to update is completed but a change is still pending,
resulting in the nodes becoming 'unready' post script completion.

This results in failed CI runs due to the cluster being in an unexpected
state.

Ensure each resource is successfully applied.

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>

add updated wait to client

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>