Avoid restarting AI if possible
In case you are deploying multiple clusters at the same time, you
do not want to restart AI because it might be in use at that
instance. Instead, first check if AI is already running and only
retstart it if the config map or persistent pod map has changed.

As a side effect, this also shaves off a few seconds of runs.