aicli: Use api_vips when interacting with aicli.
Using 'api_vip' is not supported anymore, since:
https://github.com/karmab/aicli/commit/a49b08d6348c6a8e4b9a4c7fb2388db593953d7d

This also means we need to bump the assisted-installer version to one
that works fine with recent aicli versions.  This commit bumps it to
v2.29.0 which is currently the only one compatible with aicli.

NOTE: the AI API doesn't come up if the installer and agent versions are
the ones used by SAAS in v2.29.0.  It does work with the ones from
v2.27.0; use those instead.

Validate input config and make sure it doesn't contain no longer
supported configurations.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>