clusterDeployer: Properly restrict which hosts dnsmasq gives IPs to.
Commit 8cbc29def157 ("Restrict dynamic dhcp range") tried to do that by
using the last part of the cluster IP subnet but that doesn't always
work.  There's still a chance that another host (not part of the cluster
config) requests an IP in the high part of the dhcp range.  In that case
dnsmasq was happily granting it even though that IP was to be used by a
future OCP host.

In order to properly fix this, we now configure dnsmasq to only give out
IPs to hosts whose MAC address is known.

Fixes: 8cbc29def157 ("Restrict dynamic dhcp range")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>