Fix using different network_api_port per host.
Nodes are sometimes installed with different version of OSes
resulting in different naming for the interfaces for some hosts.
This is usually configured by setting a default network_api_port
for the cluster and overwriting it for the specified hosts.
Without the patch script fails as trying to pass
network_api_port parameter twice to __init__.

Backtrace:
  File "./cluster-deployment-automation/cda.py", line 88, in <module>
    main()
  File "./cluster-deployment-automation/cda.py", line 82, in main
    main_deploy(args)
  File "./cluster-deployment-automation/cda.py", line 17, in main_deploy
    cc = ClustersConfig(args.config, args.worker_range)
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "./cluster-deployment-automation/clustersConfig.py", line 227, in __init__
    self.hosts.append(HostConfig(self.network_api_port, **e))
                      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
TypeError: HostConfig.__init__() got multiple values for argument 'network_api_port'

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>