github: don't run lint test with Python 3.9,3.10
Python 3.11 or newer is required. The README also says that since commit
0cb1b667a146 ('update docs to 3.11').

Don't run the linters against older Python versions. It seems not useful
but waste resources.

One of the linters is python-black. It's anyway unclear, why we would
run the formatter with different Python versions. But that is unchanged
for now.