test_env.sh: add script for installing python dependencies for development
It's useful to setup the machine with all pip dependencies. For example,
that is useful to have mypy working on the developer machine.

Granted, I wouldn't want to install those pip modules for my user, but
rather have a virtual environment for that.

Add a script test_venv.sh which creates a venv and installs those pip
dependencies.