host: always add a PasswordLogin for ssh-agent
Even if we have no password at hand and did not detect any ssh-keys, it
is quite possible that a SSH login would succeed. For example, if the user
adds a suitable key to ssh-agent. We don't know, before we try.
Assume we could connect and always add a PasswordLogin.

  $ # setup ssh-agent so that password-less login works.
  $ REMOTE=... ;
  $ python -c 'if 1:
        import host;
        h = host.RemoteHost("'"$REMOTE"'");
        h.ssh_connect("root");
        r = h.run("whoami ; hostname");
        print(f"result: r={r.returncode};\nout=>>>\n{r.out}<<<\nerr=>>>\n{r.err}<<<\n")
    '