host: rework _run_local() to use subprocess.run()
The code previously tried to read all of stdout. That means, if the
process writes a lot to stderr, then the socket will fill up and the
process will block and hang. We would need to read both file descriptors
at the same time.

There is already subprocess.run() for getting this all correct. Use it.