host: let Host.run() use a shell with _run_local()
Host.run() should transparently abstract whether the command is run
locally or not.

Note that for remote hosts, Host._run_remote() already interprets the
"cmd" argument as the script passed to the user's login shell.

While there are arguments for both, and while it also could be made
configurable (via a "shell" argument), it's important they they are
behaving similar.

Note that there are already various callers that rely on remote call to
be a shell scripts:

    ret = rh.run(f"rpm -qa | grep {package}")
    ...
    cmd = f"curl -k {e} --create-dirs > {wd}/{fn}"
    rh.run(cmd)

Since having a shell is more useful than not, and since the performance
overhead of going through `sh -c` is not relevant for CDA, take the
argument to be a shell script.