host: let Host.run() always use a shell
Host.run() should abstract whether the command is run locally or not.

For the remote call via ssh, we cannot run arbitrary programs. All we
can do, is construct a shell script that is passed to the login shell.
Hopefully, the login shell is a suitable POSIX shell, so we can somewhat
get a consistent behavior.

We need to make a choice, whether the cmd string is a shell script or a
argv array for execvpe(). For _run_remote() we only can have a shell
script, so restricting this would be extra work (for having something
less flexible). Also, performance is not a concern here, and we can
afford to always go through a shell.

To get consistent behavior, define that the argument is a shell script.