host: add "cwd" argument for executing command
Some callers currently call os.setcwd() before running a local command.
That is ugly, because they need to take care to reset the directory
before leaving. Also, the cwd is process wide, and we wouldn't want to
change this for other threads. Most importantly, that approach does not
work with remote shells, and optimally Host can abstract whether we run
locally or not.

Add a "cwd" argument, to set the working directory of the command. For
Host._run_remote(), we add a "cd" command to the shell script.

  $ REMOTE=...
  $ python -c 'if 1:
        import host;
        h = host.RemoteHost("'"$REMOTE"'");
        if not h.is_localhost():
          h.ssh_connect("root");
        r = h.run("whoami\nhostname\npwd", cwd="/usr/lib");
        print(f"result: r={r.returncode};\nout=>>>\n{r.out}<<<\nerr=>>>\n{r.err}<<<\n")
    '