host: improve wrapping command with sudo
"cmd" is now a shell script. Just prepending "sudo" is not correct.
Instead, we need to go through `sudo sh -c`.

Note that this breaks, if the /etc/sudoers allows certain commands.  The
sudo command now always sees "sh -c ..." and you can no longer allow
specific commands in sudoers. But CDA has an unknown, open ended list of
commands it makes. I don't think it's feasible to allow only certain
commands in sudoers, so that use case is not supported. The benefit is
that this works with cwd and env_extra arguments and with cmd being a
shell script.