host: accept argv list in Host.run()
Previously, argv had to be a shell script, which is passed via ssh.

In most cases, it's of course just a simple command with arguments. To
ensure that arguments are properly quoted, we would have to do

  rsh.run(f"ls {shlex.quote(dir)}")

As alternative, accept an argv list, which will be joined (and properly
quoted).

  rsh.run(["ls", dir])