host/test: add unit test for testing Host.run()
The test tries to ssh into 127.0.0.1. If that doesn't work, the test
will be marked as skipped.

Unfortunately, this "unit" test requires an entire machine it can SSH
into. As such, it won't work out of the box. It is still for manual
testing to ensure that the code works.

You probably need to setup ssh-agent and password-less login.
Maybe you also need a

  username ALL =(root) NOPASSWD:ALL

in /etc/sudoers.d/test-cda