clusterDeployer: Separate bridge configuration from network linking logic.
We need virtual bridges to be configured as early as possible but we don't
really want the physical network to be plugged into the bridge until at
least a minimal OCP cluster (3 masters) is up.

That's to avoid AI detecting other hosts on the network and using them
as masters.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>