assistedInstaller: download kubeadmin-password together with kubeconfig
The kubeadmin secret is useful. With it we can `oc login` and get a
token with `oc whoami -t`. I think from the X.509 certificate we cannot
get a comparable token.

Note that there is a change in behavior here. AssistedInstaller's
download_kubeconfig() does not allow to fully specify the file name. It
only allows to specify the path, and the file name is always
f"{path}/kubeconfig.{name}". This means, if the cluster.yaml previously
specified a self._cc.kubeconfig with another file name, then the file
name did not match. The user *had* to configure the expected name,
otherwise, CDA would fail to find the file under the expected location.
Fix that, by renaming the downloaded file to the requested name.

On the other hand, the resulting kubeadmin password is always in the
file f"{path}/kubeadmin-password.{name}". There is no configuration
option to change that and such flexibility seems not necessary.