assistedInstaller: download kubeadmin-password together with kubeconfig
The kubeadmin secret is useful. With it we can `oc login` and get a
token with `oc whoami -t`. I think from the X.509 certificate we cannot
get a comparable token.

Note that there is a change in behavior here. AssistedInstaller's
download_kubeconfig() does not allow to fully specify the file name. It
only allows to specify the path, and the file name is always
f"{path}/kubeconfig.{name}".  This means, previously the file name of
self._cc.kubeconfig was not honored. Now, if the file name differs, we
rename the file so that it matches what was requested.

On the other hand, the resulting kubeadmin password is always in the
file f"{path}/kubeadmin-password.{name}". The user cannot affect that
via configuration.