Support non consecutive static ip for nodes.
cluster_ip_range was set to 192.168.122.1 -> 192.168.122.n_nodes.
Configuration error was thrown if a node was configured with an
ip outside of that range.

Since we can now tear down some workers, we might end up with a cluster
with non consecutive ip. Afterwards, and without this patch, we could
not tear down workers anymore.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>