clusterNode: Properly detect nodes that failed to boot up.
If a node can't boot up for whatever reason (e.g., for VM nodes the
physical host was missing the right packages) CDA was infinitely looping
waiting for the host to come up.

Instead, detect that the boot-time future has completed and report
errors in the health_check() stage.

With this patch applied I now get (in my invalid deployment):
  2024-04-03 08:24:52 ERROR: Failed to provision node dceara-bm-worker-1: ERROR    Host does not support any virtualization options
  The program exited via sys.exit(). Exit status: -1

Signed-off-by: Dumitru Ceara <dceara@redhat.com>