common: add atomic_write() helper for writing files
The "proper" way to write small configuration files is by writing them
fully to a temporary file and do an atomic replace. Add a helper
function for that.

It's probably not perfect. It could work badly if you set certain user,
group or mode, which prevent the function from completing. It might also
not properly leave the right SELinux label (the proper solution might be
that the temporary files get the appropriate label automatically). But
especially because this isn't trivial, there is now one helper function
that can attempt to get this right.