common: add @strict_dataclass decorator for validating @dataclass
@dataclass does not enforce or check that attributes have the expected
type. Usually, that is fine.

However, ipa_to_entries() receives untrusted JSON input, parses the
fields and creates a dataclass from it. We should validate, that the
parsed arguments are as expected. As there are several fields, that gets
cumbersome to do manually. Instead, add @strict_dataclass for validating
a dataclass based on it's type information.

Somewhat surprisingly, there is no built-in method that can validate a
value against a type annotation. Hence, implement a naive version. It
doesn't need to support all types, just the ones for which we have data
classes.

See-also: https://stackoverflow.com/questions/50563546/validating-detailed-types-in-python-dataclasses