common: ignore invalid input for ipa_to_entries(), ipr_to_entries()
ipa_to_entries() and ipr_to_entries() are called like
ipa_to_entries(ipa(host)).

In this case, ipa() makes a remote call (via ssh) to `ip -json
addr|route` and passes on the output.

There is no checking that the remote call succeeds. Maybe the machine is
(temporarily) not reachable. If the machine is unreachable, then ipa()
will not return a suitable JSON output, and ipa_to_entries() will raise
an exception that leads to a crash.

If ipa() really wanted to assert that the remote call never fails, then
it should use host.run_or_die(). But it doesn't. Instead, we should
somehow handle the case where the call fails.

Even if we would require that the remote call never fails, the remote
call returns untrusted data. We should validate the JSON, and somehow
handle potentially invalid data.

We could -- as previously -- just crash on any unexpected data. However,
that is not very useful. Most calls to ipa_to_entries() are made, to
check whether a certain configuration is present. Hence, treat invalid
JSON as if the data is not present and silently ignore such errors.

Silently ignoring errors can hide bugs and making them harder to find.
On the other hand, having all callers bother with error handling for
cases that is not very interesting to them, is also cumbersome. So by
default these functions don't fail. By passing "strict_parsing=True"
the caller can get an exception for errors.