host: add "sudo" option to Host.run()
Previously, we could only enable sudo for the Host instance
(and never disable it). Note that some callers explicitly did
rsh.run("sudo ...."). It seems there is a use of asking for
sudo per-run. Add an option for that.

This is actually important. Note that Host instances get interned and
re-used. So we cannot create a new instance of host.RemoteHost(ip),
instead we get the already created one. In practice we want this, to
skip the new SSH connect. Also note that we connect to our CoreOS hosts
as "core" user. If we make an operation where we know to need super user
permissions, we should not modify the global remote host by calling
rsh.need_sudo(). Instead, elevate via sudo only for that one request.