host: fix ssh_connect_looped() to not fail after first authentication error
We have a list of logins. Previously, if the first login attempt failed with
ssh_exception.AuthenticationException, the exception was re-raised without
retrying. Fix that, by instead only raise the exception if all atttempts fail
with ssh_exception.AuthenticationException.

Also, it really seems wrong to try indefinitely. Sure, we may need to
wait for a remote machine to boot, so we may want to retry for a long
time. Still, there should be some timeout. Set the default timeout to 2
hours.