common: add ip_routes() helper
Previously, we had two separate functions: common.ipr() and
common.ipr_to_entries(). The latter would parse the JSON input while
the former returns the JSON output (as string) of the remote call.

This separation is not best.

We need one function for parsing the JSON output. This was previously
called common.ipr_to_entries() and is now renamed to
common.ip_routes_parse().

Other than that, the main usage is having a host.Host variable and
getting the routes for the host. So we don't need a common.ipr() that
makes only the `ip -json route` call. The only reasonable thing you can
do with the result of common.ipr() is to parse it.  That makes this
split wrong. Instead, drop common.ipr() and add instead
common.ip_routes(). The latter  also takes a host.Host argument, but
does the parsing in the same step.

Try:

  $ python -c 'import host, common, pprint; pprint.pp(common.ip_routes(host.LocalHost()))'
  $ ip -json route | python -c 'import common, pprint, sys; pprint.pp(common.ip_routes_parse(sys.stdin.read()))'

Also drop the unused and redundant host.ipr(). The way to get IP routes
for a host is common.ip_routes().