host: don't attach the BMC to Host
It's not clear why we would attach the BMC instance to Host. They are
used entirely independently, and calls like host.boot_iso_redfish()
are just passed through.

At most, this might have a value as we attach the BMC data to the host.
But why? If you have multiple pieces of data (a Host and a BMC), then
pass them around both to all the users that need it.

The point is to simplify Host() and move code away that is independent.