clusterSnapshotter: fix boot_iso_redfish() call for {import,export}_cluster()
Note that rh.boot_iso_redfish() requires that the host.Host instance
has self._bmc set. But it has not, the "rh" instance is just created a
few lines before, without a BMC.

Note that the caching of the host.Host via host.host_instances also
doesn't safe this. The cache considers the BMC url and
host.RemoteHost(node) does not give a host instance with BMC set.

This code is probably unused, and seems broken. The relevant part here
is to get rid of Host.boot_iso_redfish() because that will be dropped
next.

Fixes: 935fe79bba8c ('refactor BMC')