common: add common.ipaddr_norm() helper to normalize IP address from string
There is ipaddress module, that could be used for validating (and normalizing)
IP address strings. However, it is inconvenient to use. For example, it does
not generically handle IPv4 and IPv6 addresses, it requires you to handle exceptions.

Add common.ipaddr_norm(), which you can give a string (or bytes), and returns the
normalized IP address (as string) or None on error.