clustersConfig: Make node IP validation non-fatal.
The goal of the check was to ensure that requested node IPs are in the
first part of the cluster IP range.  That part of the range is not
served through DHCP to other hosts that might be connected on the same
L2 network.

However, there users that depend on a DHCP server started by a CDA
deployment to serve IPs to other CDA deployments.

Turn the IP validation into a plain error log instead of a fatal one.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>