extraConfigSriov: don't wipe the "sriov-network-operator" checkout on rebuild
Previously, the code would always fetch a new git checkout, even if the
directory already existed.

That may be good or bad. The good is, we always fetch and install
current "master" branch.

However, that also only allows us to test upstream master, and not any
other checkout.

Instead, if "sriov_network_operator_local" configuration is enabled and
we rebuild the images, then also don't wipe the repo. This allows us to
prepare a local variant that we want to test instead.