common: strictly validate IPRouteAddressInfoEntry and other dataclasses
Dataclasses have type annotations. But those types are not enforced.

On the other hand, we initialize IPRouteAddressInfoEntry (and others)
with data that we received from a remote call. The data is not trusted,
and should be validated.

Add strict checking to those data classes, so we fail early on
unexpected data.

Note that the functions that create those dataclasses from JSON data
have a "strict_parsing=False" argument. By default, those functions will
ignore any error. The benefit of that is to not burden callers with
error handling for things that don't seem particularly relevant to them.
We still add strict input validate, so that we fail early on invalid
input data (and by default handle it by silently ignoring it).