logger: honor CDA_LOG_LEVEL for configuring the logging level
CDA_LOG_LEVEL is mainly useful for manual testing or unit tests, to
increase the logging level of the logging library, without having a
command line switch implemented.

With cda.py, the "-v"/"--verbosity" option sets the logging level. It
thereby overwrites CDA_LOG_LEVEL environment variable.

However, if the command line option is not specified, it's still useful
to honor the environment variable (if set).