assistedInstallerService: Properly clean up persistent volume claims.
The pod-persistent.yml contains a few persistentVolumeClaim objects
that create a couple of persistent volumes 'ai-service-data' and
'ai-db-data'.  These volumes are not part of the pod per se and not
belong directly to any of the created containers.  So, the basic
'podman pod rm -f' doesn't remove them.

Use 'podman kube down --force' to properly destroy all the resources
defined in the configuration yml.

Unfortunately, this command is not available in older podman, so
falling back to the old method in this case.  The command should work
starting with podman 4.3.

Some older versions of podman support 'kube play --down' instead, but
also supporting this version of a fallback doesn't seem to make a lot
of sense as the number of versions that support it, but doesn't support
'kube down' is relatively small.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>