logger: add option to log all debug messages to file
For regular run, the debug output is too verbose. It overwhelms the user
with information, that is usually not used.

Unless, you want to debug a problem. If you then didn't have debug
logging enable from the start, you would wish to still have those
messages somewhere.

Solve that by adding an option "-D" which always writes debug logs to
"/tmp/cda-*.log".

The filename is not made configurable, because it's a piece of
flexibility that we won't need. It's better, if we standardize on the
filename where logs can be found, and that is "/tmp/cda-*.log".