dns: fix dnsmasq "server" option to only lookup apps.{cluster}.redhat.com
Since commit [1], libvirt's dnsmasq set "--resolvconf" option to
"/etc/resolv.conf.cda-orig". The purpose was to avoid a DNS loop. That
happened because "/etc/resolv.conf" contains "nameserver 127.0.0.1", and
the local dnsmasq.service will forward all "{cluster}.redhat.com" names
to the API host (see "/etc/dnsmasq.d/cda-cluster-{cluster}.conf" file).
The API host inside libvirt then can forward those requests back to
libvirt's dnsmasq. Which forwards to the local dnsmasq.service. And so
on. The result are log messages like

  Maximum number of concurrent DNS queries reached (max: 150)

The issue was that we forwarded  all of "{cluster}.redhat.com" to the
cluster's DNS server. We only should forward "apps", "api" and "api-int"
sub domains. The proper solution is a more fine grained selection of
domains.

Revert the change [1] to libvirt's dnsmasq and fix the resolution loop for
the local dnsmasq.service.

While at it, no longer use the "cda-cluster-{cluster}.conf" file but
switch to "/etc/dnsmasq.d/servers/cda-servers.conf". The benefit of the
"--servers-file" is that it can be reloaded via SIGHUP. Another benefit is
that this file now contains only the server options, that could also be
passed to libvirt's dnsmasq instance (but currently that is not done).
The downside is that maintaining one file for all clusters requires to
parse and update the file, while previously we could use rewrite the
per-cluster file without reading it.

Also, make sure the dnsutil API is thread safe with respect to system
changes by adding some locks.

[1] https://github.com/bn222/cluster-deployment-automation/commit/03f05ec595ee9e6b09b1b3fafc43f5075ff92f24