Print proper error message when updating go version
After installing the required version of go for building the dpu
operator, the PATH for the current process will not be updated.

We could either manually hand the updated environment to every future
call to run, however this is messy and puts a burden on the caller.

A cleaner solution might be to just error out and instruct the user to
re-start the process with the proper PATH

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>