ktoolbox: include "ktoolbox" package
This contains general purpose utilities that we have control over
(written by us).

We have multiple python projects, and there is a use in sharing code
between them. We could instead have a mono-repo, or we could copy and
paste around code, or we could reimplement things (differently).
Instead, "ktoolbox" is the place where python code can be shared.

Currently "ktoolbox" does not guarantee a stable API. So you best
refer to a specifiy commit.

[1] https://github.com/thom311/ktoolbox