clusterDeployer: extract BaseDeployer as for common parent class
IsoDeployer previously subclassed ClusterDeployer. However,
ClusterDeployer is a large thing, it's __init__() initializes
more than 15 (mostly mutable) fields. Note how IsoDeployer's
constructor does not chain up the parent's __init__()? Is that a
problem? Who knows.

Subclassing is a way for reusing code. It's good, but a IsoDeployer
is-not-a ClusterDeployer in any meaningful sense. They are quite
different.

Extract a common base class BaseDeployer, which has the functionality
that is used by both. And have the subclasses chain up the
BaseDeployer's __init__().