support "kind=iso" cluster for installing RHEL on Marvell DPU
For installing RHEL on the Marvell DPU we need to use PXE boot.  Also,
we will need to access the DPU's serial console at /dev/ttyUSB0 and
/dev/ttyUSB1 on the node.

Since we already must have code running on the worker node, also expect
that the management network port of the DPU is plugged back to the host
on interface "eno4" (configurable in marvell-tools' "--dev" option).

The secondary port of the DPU expected to be connected to the
provisioning host that runs CDA. This is usually "eno2", but
configurable via "network_api_port" in cluster YAML.

To install the DPU, CDA thus runs the podman container
quay.io/wizhat/marvell-tools:latest from marvell-tools ([1]) via SSH on
the node. This is also where the majority of the logic is.

Afterwards, it proceeds in a similar way as the other "kind=iso"
cluster deployment.

Note that we have no redfish. Instead we must SSH into the worker node
to be able to run the podman container to do the PXE boot. We thus
assume that user "core" can SSH into the node. Usually this means that
you first deploy the openshift cluster and install CoreOS on the worker
node. Then you run CDA deploy to provision the Marvell DPU on the node.

For testing and development, we can set CDA_MARVELL_TOOLS_IMAGE to use
another marvell-tools container image.

[1] https://github.com/wizhaoredhat/marvell-octeon-10-tools

https://issues.redhat.com/browse/MDC-66