deployment: Extract the future from ClusterNode
Instead of waiting for installation steps in bulk wait for every
step in series per node. To achieve that make the installation
process more serial, in other words, instead of waiting for
all nodes to do something, do the in sequence and parallelize
the sequence. This also serves as preparation for installation
retry.

Signed-off-by: Ales Musil <amusil@redhat.com>